functions/checkin_validate/requirements.txt
functions-framework==3.7.0
psycopg2-binary==2.9.9
google-cloud-firestore==2.19.0
